my-rag-app/backend/pyproject.toml
[project]
name = "my-rag-app"
version = "0.1.0"
description = "A RAG microservice using Haystack and Chroma"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "farm-haystack[all]>=1.21.2",
    "chromadb>=0.4.22",
    "chroma-haystack>=0.0.1",
    "sentence-transformers>=2.2.2",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.1",
    "types-python-dotenv>=1.0.0.3"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "integration: marks tests as integration tests"
]
addopts = "--cov=app --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true 